refactor: extract AudioFileInspection.swift from AudioUploadPolicy.swift
Splits the `AudioFileInspection` result struct out of
AudioUploadPolicy.swift so that file holds only the policy struct
(constants + oversize check).

Byte-identical move of lines 50-61. Both files import Foundation.

AudioUploadPolicy.swift: 61 → 48 lines. Tests: 667.
